+++ src/main/kotlin/com/sistema/trashmap/infrastructure/repository/EnderecoRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import java.math.BigDecimal
 
 interface EnderecoRepository : JpaRepository<Endereco, Long> {
-    fun findByCoordenadasLatitudeAndCoordenadasLongitude(latitude: BigDecimal, longitude: BigDecimal): Endereco? //Retorna um endereço segundo a latitude e longitude informadas
+    fun findByCoordenadasLatitudeAndCoordenadasLongitude(latitude: Double, longitude: Double): Endereco? //Retorna um endereço segundo a latitude e longitude informadas
 }
